+++ Utility.h
@@ -14,7 +14,7 @@ class Utility {
     static std::vector<std::string> ReadFile(const std::string& fileName);
     static void Init();
     static int Random(int min, int max);
-    static float Random();
+    static double Random();
     static float pow(float base, int exp);
     static std::vector<double> Random(unsigned int count); // return a float number between 0 and 1
 };
